% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_functions.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw a name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{Name searched}

\item{the_sex}{Sex searched}
}
\value{
Returns a graph with the number of times the name was given every year since 1900
}
\description{
Draw a name
}
\examples{

}
