% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_functions.R
\name{draw_names}
\alias{draw_names}
\title{Draw names}
\usage{
draw_names(the_names)
}
\arguments{
\item{the_names}{Names searched (vector form)}
}
\value{
Returns a graph with the number of times each name was given every year since 1900
}
\description{
Draw names
}
