% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot incidents on map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number as one-dimensional numeric vector}

\item{year}{The year of interest provided as either character or numeric}
}
\description{
If there is data for a specified state.num, \code{fars_map_state} plots
incidents on a map. The canvas limit is set to greater than 900 Longitude
and 90 Latitude. If there is no data for a specified \code{state.num} the
function exits with an error.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}

}

