% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with a list of years to summarize by.}
}
\value{
A data.frame with number of accidents by years summarized by month
}
\description{
This function summarizes yearly accidents data, by month
}
\seealso{
\link{fars_read_years}
}
