% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of interest provided as either character or numeric}
}
\value{
A character vector  in the following form: \code{accident_<year>.csv.bz2}
}
\description{
Generates a filename including the provided \code{year}.
}
\examples{
\dontrun{
fars_filename <- make_filename(2015)
}
}

