% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_functions.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{My csv reader}
\usage{
my_csv_reader(folderpath)
}
\arguments{
\item{folderpath}{The path of the folder}
}
\value{
A list of all the csv files contained in the folder
}
\description{
My csv reader
}
\examples{
my_csv_reader("~/Documents/")
}
