% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_functions.R
\name{save_as_csv}
\alias{save_as_csv}
\title{save_as_csv}
\usage{
save_as_csv(dataset, filename)
}
\arguments{
\item{dataset}{A dataset, basically}
}
\value{
The full path of the csv where this dataset has been saved
}
\description{
save_as_csv
}
