% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer specifying the number of State}

\item{year}{an integer specifying the year to be added to the file name}
}
\value{
\code{fars_map_state} will return a map plot of accidents for the given
  state and year. If there are no accidents in the specified state and year, a
  notification will be provided. If an invalid state number is provided
  an error will be returned.
}
\description{
\code{fars_map_state} will plot the accidents on a map for a given state
  and year.
}

