% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract monthly data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Years of interest provided as list or vector}
}
\value{
A list object. Each element holds monthly data for one year as
 data.frame with columns `MONTH` and `year`
}
\description{
For the specified years this function returns their monthly data. If there
is no data for the specified year available, the function will return an
error.
}
\examples{
\dontrun{
fars_year_data <- fars_read_years(2013)
fars_year_data <- fars_read_years(c(2013, 2015))
}

}

