% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Years of interest provided as list or vector}
}
\value{
A data.frame that provides the amount of monthly data per year
}
\description{
For the specified years this function summarizes their data i.e., it return
how many incidients occured in each month.
}
\examples{
\dontrun{
fars_summary <- fars_summarize_years(2013)
fars_summary <- fars_summarize_years(c(2013, 2014))
}

}

