% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Observations by Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer specifying the year to be added to the file name}
}
\value{
\code{fars_summarize_years} will return data frame.
}
\description{
\code{fars_summarize_years} will read in multiple Fatality Analysis Reporting
  System data files based on the years provided and summarise the number of
  observations by year and month.
}

