% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{make_filename}
\alias{make_filename}
\title{Make File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer specifying the year to be added to the file name}
}
\value{
\code{make_file} will return a string that speficy the file name based on the year provided.
  For example, if 2016 is provided as the year, the returned name will be
  "accident_2017.csv.bz2".
}
\description{
\code{make_file} creates a name for the accident csv file based on the
  year provided.
}

